+++ tests/test_kxpos/src/test_kxpos.cpp
@@ -27,7 +27,7 @@ int main(int argc, char* argv[])
       float kx = kxPos(kl, nxLocal, nyLocal, nf);
 
       if (kx != (float)kxx) {
-         pvError().printf("FAILED:TEST_KXPOS: (k,kx) = (%d,%f)\n", kl, kx);
+         pvError().printf("FAILED:TEST_KXPOS: (k,kx) = (%d,%f)\n", kl, (double)kx);
       }
    }
 
@@ -49,7 +49,7 @@ int main(int argc, char* argv[])
       float kx = kxPos(kl, nxLocal, nyLocal, nf);
 
       if ((int)kx-kxx != 0) {
-         pvError().printf("FAILED:TEST_KXPOS: (k,kx) = (%d,%f)\n", kl, kx);
+         pvError().printf("FAILED:TEST_KXPOS: (k,kx) = (%d,%f)\n", kl, (double)kx);
       }
    }
 
@@ -72,7 +72,7 @@ int main(int argc, char* argv[])
       float kx = kxPos(kl, nxLocal, nyLocal, nf);
 
       if ((int)kx-kxx != 0) {
-         pvError().printf("FAILED:TEST_KXPOS: (k,kx) = (%d,%f)\n", kl, kx);
+         pvError().printf("FAILED:TEST_KXPOS: (k,kx) = (%d,%f)\n", kl, (double)kx);
       }
    }
 
@@ -96,7 +96,7 @@ int main(int argc, char* argv[])
       float kx = kxPos(kl, nxLocal, nyLocal, nf);
 
       if ((int)kx-kxx != 0) {
-         pvError().printf("FAILED:TEST_KXPOS: (k,kx) = (%d,%f)\n", kl, kx);
+         pvError().printf("FAILED:TEST_KXPOS: (k,kx) = (%d,%f)\n", kl, (double)kx);
       }
    }
